trucking_app/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
